Input int working! - clean up gdt

+++ src/impl/x86_64/print.c
@@ -8,20 +8,19 @@ struct Char
     uint8_t character;
     uint8_t color;
 };
-
 struct Char *buffer = (struct Char *)0xb8000;
-size_t col = 0;
-size_t row = 0;
+int col = 0;
+int row = 0;
 uint8_t color = PRINT_COLOR_WHITE | PRINT_COLOR_BLACK << 4;
 
-void clear_row(size_t row)
+void clear_row(int row)
 {
     struct Char empty = (struct Char){
         character : ' ',
         color : color,
     };
 
-    for (size_t col = 0; col < NUM_COLS; col++)
+    for (int col = 0; col < NUM_COLS; col++)
     {
         buffer[col + NUM_COLS * row] = empty;
     }
